- better package structure - interface is not necessary here - extract one if needed
