+++ src/org/intellij/ideaplugins/tabswitch/action/NextTabAction.java
@@ -15,8 +15,8 @@
  */
 package org.intellij.ideaplugins.tabswitch.action;
 
-import org.intellij.ideaplugins.tabswitch.filefetchers.FileFetcher;
-import org.intellij.ideaplugins.tabswitch.filefetchers.FileFetcherOpenTabFiles;
+import org.intellij.ideaplugins.tabswitch.filefetcher.FileFetcher;
+import org.intellij.ideaplugins.tabswitch.filefetcher.FileFetcherOpenTabFiles;
 
 import com.intellij.openapi.vfs.VirtualFile;
