+++ src/org/intellij/ideaplugins/tabswitch/action/NextVcsChangeAction.java
@@ -1,7 +1,7 @@
 package org.intellij.ideaplugins.tabswitch.action;
 
-import org.intellij.ideaplugins.tabswitch.filefetchers.FileFetcher;
-import org.intellij.ideaplugins.tabswitch.filefetchers.FileFetcherChangedFilesInVcs;
+import org.intellij.ideaplugins.tabswitch.filefetcher.FileFetcher;
+import org.intellij.ideaplugins.tabswitch.filefetcher.FileFetcherChangedFilesInVcs;
 
 import com.intellij.openapi.vfs.VirtualFile;
