+++ src/org/intellij/ideaplugins/tabswitch/component/FooterComponentFactory.java
@@ -1,4 +1,4 @@
-package org.intellij.ideaplugins.tabswitch;
+package org.intellij.ideaplugins.tabswitch.component;
 
 import java.awt.BorderLayout;
 import java.awt.Graphics;
