+++ src/org/intellij/ideaplugins/tabswitch/component/PathLabelComponentFactory.java
@@ -1,4 +1,4 @@
-package org.intellij.ideaplugins.tabswitch;
+package org.intellij.ideaplugins.tabswitch.component;
 
 import javax.swing.JLabel;
 import javax.swing.SwingConstants;
